Android: adjust app permissions
- AGS Player needs access to external storage
- Single Game doesn't need this access
